refactor pt3
