+++ gpt2giga/openapi_specs/__init__.py
@@ -1,25 +1,1 @@
-"""Public OpenAPI schema helpers."""
-
-from gpt2giga.openapi_specs.anthropic import (
-    anthropic_count_tokens_openapi_extra,
-    anthropic_messages_openapi_extra,
-)
-from gpt2giga.openapi_specs.api import (
-    batches_openapi_extra,
-    chat_completions_openapi_extra,
-    embeddings_openapi_extra,
-    files_openapi_extra,
-    responses_openapi_extra,
-)
-from gpt2giga.openapi_specs.common import _request_body_oneof
-
-__all__ = [
-    "_request_body_oneof",
-    "anthropic_count_tokens_openapi_extra",
-    "anthropic_messages_openapi_extra",
-    "batches_openapi_extra",
-    "chat_completions_openapi_extra",
-    "embeddings_openapi_extra",
-    "files_openapi_extra",
-    "responses_openapi_extra",
-]
+"""OpenAPI schema helper modules."""
