+++ gpt2giga/protocol/anthropic/__init__.py
@@ -1,0 +1,1 @@
+"""Anthropic protocol translation modules."""
