+++ gpt2giga/routers/__init__.py
@@ -1,12 +1,1 @@
-from .anthropic import router as anthropic_router
-from .api import router as api_router
-from .logs_router import logs_api_router, logs_router
-from .system_router import system_router
-
-__all__ = [
-    "anthropic_router",
-    "api_router",
-    "logs_api_router",
-    "logs_router",
-    "system_router",
-]
+"""Router packages and modules."""
