+++ gpt2giga/routers/anthropic/messages.py
@@ -5,22 +5,22 @@ from typing import Dict, List
 from fastapi import APIRouter, Request
 from fastapi.responses import StreamingResponse
 
+from gpt2giga.app_state import get_gigachat_client
 from gpt2giga.common.exceptions import exceptions_handler
 from gpt2giga.common.request_json import read_request_json
 from gpt2giga.logger import rquid_context
-from gpt2giga.openapi_specs import (
+from gpt2giga.openapi_specs.anthropic import (
     anthropic_count_tokens_openapi_extra,
     anthropic_messages_openapi_extra,
 )
-from gpt2giga.routers.anthropic.conversion import (
+from gpt2giga.protocol.anthropic.request import (
     _build_openai_data_from_anthropic_request,
     _convert_anthropic_messages_to_openai,
     _extract_text_from_openai_messages,
     _extract_tool_definitions_text,
 )
-from gpt2giga.routers.anthropic.responses import _build_anthropic_response
-from gpt2giga.routers.anthropic.streaming import _stream_anthropic_generator
-from gpt2giga.routers.state import get_gigachat_client
+from gpt2giga.protocol.anthropic.response import _build_anthropic_response
+from gpt2giga.protocol.anthropic.streaming import _stream_anthropic_generator
 
 router = APIRouter(tags=["Anthropic"])
