+++ gpt2giga/routers/openai/helpers.py
@@ -4,8 +4,9 @@ from typing import Optional
 
 from fastapi import Request
 
+from gpt2giga.app_state import get_batch_store, get_gigachat_client
 from gpt2giga.protocol.batches import infer_openai_file_purpose
-from gpt2giga.routers.state import get_batch_store, get_gigachat_client
+from gpt2giga.common.tools import convert_tool_to_giga_functions
 
 
 def _paginate_items(
@@ -39,6 +40,15 @@ def _serialize_file_object(file_obj, stored_metadata: Optional[dict] = None) -> 
         "expires_at": stored_metadata.get("expires_at"),
         "status_details": stored_metadata.get("status_details"),
     }
+
+
+def populate_giga_functions(data: dict, logger) -> None:
+    """Populate GigaChat-compatible function definitions when tools are present."""
+    if "tools" not in data and "functions" not in data:
+        return
+    data["functions"] = convert_tool_to_giga_functions(data)
+    if logger:
+        logger.debug(f"Functions count: {len(data['functions'])}")
 
 
 async def _load_batch_output_content(request: Request, file_id: str) -> bytes:
