+++ gpt2giga/routers/openai/models.py
@@ -7,10 +7,10 @@ from fastapi import APIRouter, Request
 from openai.pagination import AsyncPage
 from openai.types import Model as OpenAIModel
 
+from gpt2giga.app_state import get_gigachat_client
 from gpt2giga.common.exceptions import exceptions_handler
-from gpt2giga.routers.state import get_gigachat_client
 
-router = APIRouter(tags=["API"])
+router = APIRouter(tags=["OpenAI"])
 
 
 @router.get("/models")
