+++ tests/test_router/test_router_stream_chat.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 
 from gpt2giga.models.config import ProxyConfig
 from gpt2giga.protocol import ResponseProcessor
-from gpt2giga.routers.api import router
+from gpt2giga.routers.openai import router
 
 
 class FakeGigachat:
